subtitles: Fix for characters such as tilde, single and double quotes not displayed in subtitles See: https://github.com/popcornmix/omxplayer/issues/303 Thanks to vinv for discovery

+++ SubtitleRenderer.cpp
@@ -159,7 +159,7 @@ void SubtitleRenderer::load_glyph(InternalChar ch) {
         image = softened_image;
 
         glyph_origin[0] = static_cast<VGfloat>(padding - bit_glyph->left);
-        glyph_origin[1] = static_cast<VGfloat>(padding + bitmap.rows - bit_glyph->top - 1);
+        glyph_origin[1] = static_cast<VGfloat>(padding + static_cast<int>(bitmap.rows) - bit_glyph->top - 1);
       }
 
       escapement[0] = static_cast<VGfloat>((ft_face->glyph->advance.x + 32) / 64);
